update end point create and delete address

+++ domains/Purchasing/Business/PurchaseOrder.ts
@@ -74,26 +74,34 @@ export class PurchaseOrder {
     }
 
 
-    public static async createAddress(ctx, datas){
+    public static async createAddress(ctx, id, datas){
         let customerId = SecurityService.getCurrentCustomerId(ctx);
-        let purchasingService = await IntegrationService.getPurchasingService(customerId);
-        if (!purchasingService) {
-            throw ErrorFactory
-                .createError(`Customer has not installed any purchasing app yet`,400,'INVALID_SERVICE');
-        }
+        if(customerId == id) {
+            let purchasingService = await IntegrationService.getPurchasingService(customerId);
+            if (!purchasingService) {
+                throw ErrorFactory
+                    .createError(`Customer has not installed any purchasing app yet`,400,'INVALID_SERVICE');
+            }
 
-        return await purchasingService.createAddress(customerId, datas);
+            return await purchasingService.createAddress(customerId, datas);
+        }
+        throw ErrorFactory
+            .createError(`Customer id invalid`, 401, 'INVALID_ID');
     }
 
     public static async deleteAddress(ctx, id){
         let customerId = SecurityService.getCurrentCustomerId(ctx);
-        let purchasingService = await IntegrationService.getPurchasingService(customerId);
-        if (!purchasingService) {
-            throw ErrorFactory
-                .createError(`Customer has not installed any purchasing app yet`,400,'INVALID_SERVICE');
-        }
+        if(customerId == id) {
+            let purchasingService = await IntegrationService.getPurchasingService(customerId);
+            if (!purchasingService) {
+                throw ErrorFactory
+                    .createError(`Customer has not installed any purchasing app yet`,400,'INVALID_SERVICE');
+            }
 
-        return await purchasingService.deleteAddress(customerId, id);
+            return await purchasingService.deleteAddress(customerId, id);
+        }
+        throw ErrorFactory
+            .createError(`Customer id invalid`, 401, 'INVALID_ID');
     }
 
 }
